Move the code a bit
